[IMP] osi_analytic_segments_defaults: Default value for segment one
