+++ osi_analytic_segments_defaults/models/__init__.py
@@ -1,5 +1,7 @@
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
-
-from . import account_analytic_default
-from . import purchase_order
-from . import sale_order_line
+from . import (
+    account_analytic_default,
+    account_invoice_line,
+    purchase_order,
+    sale_order_line,
+)
